API Reference

As you will see in the Code Outline and getting started, most of the code in this package revolves around creating instances of model, which have been formatted to be solved. There are some convenience functions for creating these structs. There are also some functions for parsing solutions. another change. another change.

Airfoils

DynamicStallModels.Airfoil — Type

Airfoil(polar, cl, cd, cm, dcldalpha, alpha0, alphasep, A, b, T)

Inputs

polar::Array{TF, 2} - A matrix of floats describing the airfoil polar. This includes the angle of attack (radians), and coefficients of lift, drag, and moment.

cl::Tfit - A fit of the coefficient of lift as a function of the angle of attack (radians).

cd::Tfit - A fit of the coefficient of drag as a function of the angle of attack (radians).

cm::Tfit - A fit of the coefficient of moment as a function of the angle of attack (radians).

dcldalpha::TF - The lift curve slope in the linear region (1/radians). Typically near 2 pi.

alpha0::TF - The zero lift angle of attack (radians).

alphasep::Array{TF, 1} - A vector of floats holding the angles of attack at which flow fully seperates from the airfoil. In order from least to greatest.

A::Array{TF, 1} - A vector of floats holding the A dynamic constants for the airfoil.

b::Array{TF, 1} - A vector of floats holding the b dynamic constants for the airfoil.

T::Array{TF, 1} - A vector of floats holding the time constants for the airfoil.

DynamicStallModels.simpleairfoil — Function

simpleairfoil(polar)

A function that takes a simple airfoil polar to make a dynamic airfoil. This function makes some gross assumptions.

Inputs

polar::Array{TF, 2} - A simple airfoil polar including alpha, lift, and drag.

Outputs

Airfoil

Notes

The function assumes that the lift curve slope is 2 pi, that the zero lift angle of attack is zero, and uses AeroDyn's default dynamic airfoil coefficients. The seperation angles of attack are the angles of attack at the minimum and maximum lift.

DynamicStallModels.airfoil — Function

airfoil(polar; A = [0.3, 0.7], b = [0.14, 0.53], T = [1.7, 3.0])

A slightly more complex version of simpleairfoil. Takes a polar and numberically finds some characteristics.

Inputs

polar::Array{TF, 2}

A::Array{TF, 1}

b::Array{TF, 1}

T::Array{TF, 1}

Outputs

Airfoil

Risø

DynamicStallModels.Riso — Type

Riso(detype, n, airfoils)