API Reference

As you will see in the Code Outline and getting started, most of the code in this package revolves around creating instances of model, which have been formatted to be solved. There are some convenience functions for creating these structs. There are also some functions for parsing solutions. another change. another change.

Airfoils

DynamicStallModels.Airfoil — Type

Airfoil(polar::Array{TF, 2}, cl::Tfit, cd::Tfit, cm::Tfit, dcldalpha::TF, alpha0::TF, alphasep::Array{TF, 1}, A::Array{TF, 1}, b::Array{TF, 1}, T::Array{TF, 1})

A struct to hold all of the airfoil polars, fits, and dynamic coefficients.

Inputs

polar - A matrix of floats describing the airfoil polar. This includes the angle of attack (radians), and coefficients of lift, drag, and moment.

cl - A fit of the coefficient of lift as a function of the angle of attack (radians).

cd - A fit of the coefficient of drag as a function of the angle of attack (radians).

cm - A fit of the coefficient of moment as a function of the angle of attack (radians).

dcldalpha - The lift curve slope in the linear region (1/radians). Typically near 2 pi.

alpha0 - The zero lift angle of attack (radians).

alphasep - A vector of floats holding the angles of attack at which flow fully seperates from the airfoil. In order from least to greatest.

A - A vector of floats holding the A dynamic constants for the airfoil.

b - A vector of floats holding the b dynamic constants for the airfoil.

T - A vector of floats holding the time constants for the airfoil.

S - A vector of floats holding the S constants are best fit constants for the separation point curve.

s - A fit of the the separation point curve.

Missing docstring.

Missing docstring for simpleairfoil. Check Documenter's build log for details.

Missing docstring.

Missing docstring for airfoil. Check Documenter's build log for details.

Beddoes-Leishman

DynamicStallModels.BeddoesLeishman — Type

BeddoesLeishman(detype::DEType, n::Int, airfoils::Array{Airfoil, 1}, version::Int)

The Beddoes-Leishman model struct. It stores airfoil data for every section to be simulated. It can be used as a method to return updated states or state rates depending on it's DEType.

Inputs

detype - The type of model it is, Continuous(), Discrete(), or Discrete().

n - The number of sections to be simulated.

airfoils - A vector of Airfoil structs, one corresponding to each section to be simulated.

version - Which version of the Discrete implementation. 1) original. 2) AeroDyn original. 3) AeroDyn Gonzalez. 4) AeroDyn Minema

Larsen

Missing docstring.

Missing docstring for Larsen. Check Documenter's build log for details.

Onera

Missing docstring.

Missing docstring for Onera. Check Documenter's build log for details.

Oye

DynamicStallModels.Oye — Type

Oye(detype::DEType, n::Int, airfoils::Array{Airfoil, 1})

The Øye model struct. It stores airfoil data for every section to be simulated. It can be used as a method to return updated states or state rates depending on it's DEType.

Inputs

detype - The type of model it is, Continuous() or Discrete().

cflag::Int - A flag to apply the separation delay to the coefficient of 1) lift, 2) normal force.

version::Int - A flag to say whether to use 1) Hansen 2008, or 2) Larsen's Hermite interpolation from Faber's 2018's implementation of the model.

A::Float - Dynamic stall coefficient.

Risø

DynamicStallModels.Riso — Type

Riso(detype::DEType, n::Int, airfoils::Array{Airfoil, 1})

The Risø model struct. It stores airfoil data for every section to be simulated. It can be used as a method to return updated states or state rates depending on it's DEType.

Inputs

detype - The type of model it is, Continuous() or Discrete().

n - The number of sections to be simulated.

airfoils - A vector of Airfoil structs, one corresponding to each section to be simulated.

Missing docstring.

Missing docstring for riso. Check Documenter's build log for details.