API Reference

As you will see in the Code Outline and getting started, most of the code in this package revolves around creating instances of model, which have been formatted to be solved. There are some convenience functions for creating these structs. There are also some functions for parsing solutions. another change. another change.

Airfoils

DynamicStallModels.Airfoil — Type

Airfoil(polar::Array{TF, 2}, cl::Tfit, cd::Tfit, cm::Tfit, dcldalpha::TF, alpha0::TF, alphasep::Array{TF, 1}, A::Array{TF, 1}, b::Array{TF, 1}, T::Array{TF, 1})

A struct to hold all of the airfoil polars, fits, and dynamic coefficients.

Inputs

polar - A matrix of floats describing the airfoil polar. This includes the angle of attack (radians), and coefficients of lift, drag, and moment.

cl - A fit of the coefficient of lift as a function of the angle of attack (radians).

cd - A fit of the coefficient of drag as a function of the angle of attack (radians).

cm - A fit of the coefficient of moment as a function of the angle of attack (radians).

dcldalpha - The lift curve slope in the linear region (1/radians). Typically near 2 pi.

alpha0 - The zero lift angle of attack (radians).

alphasep - A vector of floats holding the angles of attack at which flow fully seperates from the airfoil. In order from least to greatest.

A - A vector of floats holding the A dynamic constants for the airfoil.

b - A vector of floats holding the b dynamic constants for the airfoil.

T - A vector of floats holding the time constants for the airfoil.

DynamicStallModels.simpleairfoil — Function

simpleairfoil(polar::Array{TF, 2})

A function that takes a simple airfoil polar to make a dynamic airfoil. The function assumes that the lift curve slope is 2 pi, that the zero lift angle of attack is zero, and uses AeroDyn's default dynamic airfoil coefficients. The seperation angles of attack are the angles of attack at the minimum and maximum lift.

Inputs

polar - A simple airfoil polar including alpha, lift, and drag.

Outputs

Airfoil

DynamicStallModels.airfoil — Function

airfoil(polar::Array{TF, 2}; A::Array{TF, 1} = [0.3, 0.7], b::Array{TF, 1} = [0.14, 0.53], T::Array{TF, 1} = [1.7, 3.0])

A slightly more complex version of simpleairfoil. Takes a polar and numerically finds some characteristics.

Inputs

polar - A matrix of floats describing the airfoil polar. This includes the angle of attack (radians), and coefficients of lift, drag, and moment.

A - A vector of floats holding the A dynamic constants for the airfoil.

b - A vector of floats holding the b dynamic constants for the airfoil.

T - A vector of floats holding the time constants for the airfoil.

Outputs

Airfoil

Beddoes-Leishman

DynamicStallModels.BeddoesLeishman — Type

BeddoesLeishman(detype::DEType, n::Int, airfoils::Array{Airfoil, 1})

The Beddoes-Leishman model struct. It stores airfoil data for every section to be simulated. It can be used as a method to return updated states or state rates depending on it's DEType.

Inputs

detype - The type of model it is, Functional(), Iterative(), or Indicial().

n - The number of sections to be simulated.

airfoils - A vector of Airfoil structs, one corresponding to each section to be simulated.

Larsen

DynamicStallModels.Larsen — Type

Larsen(detype::DEType, n::Int, airfoils::Array{Airfoil, 1})

The Larsen model struct. It stores airfoil data for every section to be simulated. It can be used as a method to return updated states or state rates depending on it's DEType.

Inputs

detype - The type of model it is, Functional(), Iterative(), or Indicial().

n - The number of sections to be simulated.

airfoils - A vector of Airfoil structs, one corresponding to each section to be simulated.

Onera

DynamicStallModels.Onera — Type

Onera(detype::DEType, n::Int, airfoils::Array{Airfoil, 1})

The Onera model struct. It stores airfoil data for every section to be simulated. It can be used as a method to return updated states or state rates depending on it's DEType.

Inputs

detype - The type of model it is, Functional(), Iterative(), or Indicial().

n - The number of sections to be simulated.

airfoils - A vector of Airfoil structs, one corresponding to each section to be simulated.

Oye

DynamicStallModels.Oye — Type

Oye(detype::DEType, n::Int, airfoils::Array{Airfoil, 1})

The Øye model struct. It stores airfoil data for every section to be simulated. It can be used as a method to return updated states or state rates depending on it's DEType.

Inputs

detype - The type of model it is, Functional(), Iterative(), or Indicial().

n - The number of sections to be simulated.

airfoils - A vector of Airfoil structs, one corresponding to each section to be simulated.

Risø

DynamicStallModels.Riso — Type

Riso(detype::DEType, n::Int, airfoils::Array{Airfoil, 1})

The Risø model struct. It stores airfoil data for every section to be simulated. It can be used as a method to return updated states or state rates depending on it's DEType.

Inputs

detype - The type of model it is, Functional(), Iterative(), or Indicial().

n - The number of sections to be simulated.

airfoils - A vector of Airfoil structs, one corresponding to each section to be simulated.

DynamicStallModels.riso — Function

riso(airfoils::Array{Airfoil, 1}; detype::DEType=Iterative())

A convenience. constructor for the Risø model.

Inputs

airfoils - A vector of Airfoil structs, one corresponding to each section to be simulated.

detype - The DEType of the model. Defaults to Iterative().